requirements.txt
torch>=2.0.0
marker-pdf>=1.0.0
Pillow>=9.0.0  # for image handling
